+++ front/babel.config.js
@@ -4,18 +4,12 @@ module.exports = {
       targets: {
         browsers: ['last 3 versions', '> 2%', 'ie >= 10', 'Firefox >= 30', 'Chrome >= 30']
       },
-      modules: false,
-      loose: true,
-      useBuiltIns: 'entry'
+      modules: 'commonjs',
+      useBuiltIns: 'usage'
     }]
   ],
   plugins: [
     '@babel/plugin-transform-runtime',
     '@babel/plugin-syntax-dynamic-import'
-  ],
-  env: {
-    test: {
-      plugins: ['transform-es2015-modules-commonjs']
-    }
-  }
+  ]
 }
