+++ front/client/app.ts
@@ -1,16 +1,10 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 import Vuex from 'vuex'
-import Vuetify from 'vuetify'
+import vuetify from 'client/plugins/vuetify'
 import { sync } from 'vuex-router-sync'
 import ALink from 'client/components/ALink.vue'
 import confirmDialogMixin from 'client/mixins/confirmDialogMixin'
-import 'vuetify/dist/vuetify.css'
-import { library } from '@fortawesome/fontawesome-svg-core'
-import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
-import { fas } from '@fortawesome/free-solid-svg-icons'
-import { faSquare } from '@fortawesome/free-regular-svg-icons'
-library.add(fas, faSquare)
 
 import 'client/assets/code-highlight.css'
 
@@ -18,27 +12,11 @@ import AppComponent from './App.vue'
 import { createRouter } from './router'
 import { createStore } from './store'
 
-const icons = {}
-// TODO import and add only used icons for production
-Object.values({ ...fas, faSquare }).forEach(({ prefix, iconName }) => {
-  icons[iconName] = {
-    component: 'font-awesome-icon',
-    props: {
-      icon: [prefix, iconName]
-    }
-  }
-})
-
 function initVue () {
   Vue.use(VueRouter)
   Vue.use(Vuex)
   Vue.mixin(confirmDialogMixin)
   Vue.component('ALink', ALink)
-  Vue.component('font-awesome-icon', FontAwesomeIcon)
-  Vue.use(Vuetify, {
-    iconfont: 'faSvg',
-    icons
-  })
 }
 
 function createApp () {
@@ -50,6 +28,7 @@ function createApp () {
   const app = new Vue({
     router,
     store,
+    vuetify,
     render: h => h(AppComponent)
   })
