+++ front/package.json
@@ -28,67 +28,73 @@
     ]
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.17",
-    "@fortawesome/free-regular-svg-icons": "^5.8.2",
-    "@fortawesome/free-solid-svg-icons": "^5.8.1",
+    "@fortawesome/fontawesome-svg-core": "^1.2.19",
+    "@fortawesome/free-regular-svg-icons": "^5.9.0",
+    "@fortawesome/free-solid-svg-icons": "^5.9.0",
     "@fortawesome/vue-fontawesome": "^0.1.6",
-    "axios": "^0.18.1",
-    "easymde": "^2.6.0",
+    "axios": "^0.19.0",
+    "easymde": "^2.7.0",
     "koa": "^2.7.0",
     "koa-bodyparser": "^4.2.1",
     "koa-mount": "^4.0.0",
     "koa-proxy": "^0.9.0",
     "koa-static": "^5.0.0",
-    "lodash": "^4.17.13",
+    "lodash": "^4.17.15",
     "moment": "^2.24.0",
     "normalize-url": "^4.3.0",
     "nprogress": "^0.2.0",
     "vue": "^2.6.10",
-    "vue-class-component": "^7.0.2",
-    "vue-mixin-decorator": "^1.1.1",
-    "vue-property-decorator": "^8.1.0",
-    "vue-router": "^3.0.4",
+    "vue-class-component": "^7.1.0",
+    "vue-mixin-decorator": "^1.2.0",
+    "vue-property-decorator": "^8.2.1",
+    "vue-router": "^3.0.7",
     "vue-server-renderer": "^2.6.10",
-    "vuetify": "^1.5.11",
-    "vuex": "^3.1.0",
+    "vuetify": "^2.0.0",
+    "vuex": "^3.1.1",
     "vuex-router-sync": "^5.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.3",
+    "fibers": "^4.0.1",
+    "@babel/core": "^7.5.5",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
+    "@babel/plugin-transform-runtime": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
+    "@babel/runtime": "^7.5.5",
     "@vue/preload-webpack-plugin": "^1.1.0",
-    "autoprefixer": "^9.5.1",
-    "babel-loader": "^8.0.5",
-    "concurrently": "^4.1.0",
-    "copyfiles": "^2.1.0",
+    "autoprefixer": "^9.6.1",
+    "babel-loader": "^8.0.6",
+    "concurrently": "^4.1.1",
+    "copyfiles": "^2.1.1",
     "cross-env": "^5.2.0",
-    "css-loader": "^2.1.1",
-    "file-loader": "^3.0.1",
-    "fork-ts-checker-webpack-plugin": "^1.0.2",
+    "css-loader": "^3.1.0",
+    "deepmerge": "^4.0.0",
+    "file-loader": "^4.1.0",
+    "fork-ts-checker-webpack-plugin": "^1.4.3",
     "friendly-errors-webpack-plugin": "^1.7.0",
-    "koa-webpack": "^5.2.2",
+    "koa-webpack": "^5.2.4",
     "memory-fs": "^0.4.1",
-    "nodemon": "^1.18.11",
+    "nodemon": "^1.19.1",
     "postcss-loader": "^3.0.0",
-    "postcss-preset-env": "^6.6.0",
+    "postcss-preset-env": "^6.7.0",
     "precss": "^4.0.0",
     "rimraf": "^2.6.3",
-    "testcafe": "^1.1.2",
+    "sass": "^1.22.7",
+    "sass-loader": "^7.1.0",
+    "testcafe": "^1.3.3",
     "testcafe-vue-selectors": "^3.1.0",
-    "ts-loader": "^5.3.3",
-    "ts-node": "^8.0.3",
-    "tslint": "^5.15.0",
-    "typescript": "^3.4.3",
-    "url-loader": "^1.1.2",
-    "vue-loader": "^15.7.0",
+    "ts-loader": "^6.0.4",
+    "ts-node": "^8.3.0",
+    "tslint": "^5.18.0",
+    "typescript": "^3.5.3",
+    "url-loader": "^2.1.0",
+    "vue-loader": "^15.7.1",
     "vue-style-loader": "^4.1.2",
     "vue-template-compiler": "^2.6.10",
-    "webpack": "^4.30.0",
-    "webpack-cli": "^3.3.0",
+    "vuetify-loader": "^1.3.0",
+    "webpack": "^4.37.0",
+    "webpack-cli": "^3.3.6",
     "webpack-merge": "^4.2.1",
     "webpack-node-externals": "^1.7.2",
-    "webpackbar": "^3.1.5"
+    "webpackbar": "^3.2.0"
   }
 }
